+++ oopsmails-common/oopsmails-common-annotation/src/main/java/com/oopsmails/common/annotation/audit/AuditArg.java
@@ -1,4 +1,4 @@
-package com.oopsmails.springboot.mockbackend.annotation.audit;
+package com.oopsmails.common.annotation.audit;
 
 public interface AuditArg {
     String getAuditArgTarget();
