+++ oopsmails-common/oopsmails-common-annotation/src/main/java/com/oopsmails/common/annotation/audit/LoggingAudit.java
@@ -1,6 +1,6 @@
-package com.oopsmails.springboot.mockbackend.annotation.audit;
+package com.oopsmails.common.annotation.audit;
 
-import com.oopsmails.springboot.mockbackend.model.logging.LoggingOrigin;
+import com.oopsmails.common.annotation.model.logging.LoggingOrigin;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -8,9 +8,12 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.METHOD, ElementType.TYPE, ElementType.PARAMETER, ElementType.FIELD}) // only METHOD now as example, will add others
+@Target({ElementType.METHOD, ElementType.TYPE, ElementType.PARAMETER, ElementType.FIELD})
+// only METHOD now as example, will add others
 public @interface LoggingAudit {
     LoggingOrigin origin();
+
     String message() default "";
+
     String type() default "";
 }
