+++ oopsmails-common/oopsmails-common-annotation/src/main/java/com/oopsmails/common/annotation/config/OopsmailsCommonAnnotationConfig.java
@@ -1,0 +1,9 @@
+package com.oopsmails.common.annotation.config;
+
+import org.springframework.boot.SpringBootConfiguration;
+import org.springframework.context.annotation.ComponentScan;
+
+@SpringBootConfiguration
+@ComponentScan("com.oopsmails.common.annotation")
+public class OopsmailsCommonAnnotationConfig {
+}
