+++ oopsmails-common/oopsmails-common-annotation/src/main/java/com/oopsmails/common/annotation/crypto/CryptoCheckPayload.java
@@ -1,4 +1,4 @@
-package com.oopsmails.springboot.mockbackend.annotation.crypto;
+package com.oopsmails.common.annotation.crypto;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -6,11 +6,8 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target({ ElementType.METHOD })
+@Target({ ElementType.PARAMETER })
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-public @interface CryptoCheck {
-
-    String cryptoSecretProperty();
-    String mandatoryProperty();
+public @interface CryptoCheckPayload {
 }
