+++ oopsmails-common/oopsmails-common-annotation/src/main/java/com/oopsmails/common/annotation/crypto/CryptoCheckSignature.java
@@ -1,4 +1,4 @@
-package com.oopsmails.springboot.mockbackend.annotation.crypto;
+package com.oopsmails.common.annotation.crypto;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -9,5 +9,5 @@ import java.lang.annotation.Target;
 @Target({ ElementType.PARAMETER })
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-public @interface CryptoCheckPayload {
+public @interface CryptoCheckSignature {
 }
