+++ oopsmails-common/oopsmails-common-annotation/src/main/java/com/oopsmails/common/annotation/model/logging/AppLogLevel.java
@@ -1,4 +1,4 @@
-package com.oopsmails.springboot.mockbackend.model.logging;
+package com.oopsmails.common.annotation.model.logging;
 
 import lombok.Getter;
