+++ oopsmails-common/oopsmails-common-annotation/src/main/java/com/oopsmails/common/annotation/model/logging/ClientType.java
@@ -1,0 +1,5 @@
+package com.oopsmails.common.annotation.model.logging;
+
+public enum ClientType {
+    POSTMAN
+}
