+++ oopsmails-common/oopsmails-common-domain/src/main/java/com/oopsmails/common/constant/CommonLoggingConstants.java
@@ -1,4 +1,4 @@
-package com.oopsmails.common.filter.util;
+package com.oopsmails.common.constant;
 
 public class CommonLoggingConstants {
     public static final String REQUEST_HEADER_CORRELATION_ID = "correlation_id";
