+++ oopsmails-common/oopsmails-common-filter/src/main/java/com/oopsmails/common/filter/CommonRequestFilter.java
@@ -1,6 +1,6 @@
 package com.oopsmails.common.filter;
 
-import com.oopsmails.common.filter.util.CommonLoggingConstants;
+import com.oopsmails.common.constant.CommonLoggingConstants;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.MDC;
