+++ oopsmails-common/pom.xml
@@ -18,6 +18,7 @@
     <modules>
         <module>oopsmails-common-domain</module>
         <module>oopsmails-common-filter</module>
+        <module>oopsmails-common-annotation</module>
     </modules>
 
     <properties>
@@ -27,6 +28,8 @@
         <commons-beanutils.version>1.9.3</commons-beanutils.version>
         <commons-io.version>2.6</commons-io.version>
         <jackson.core.version>2.11.4</jackson.core.version>
+        <aspectjweaver.version>1.9.6</aspectjweaver.version>
+        <commons-codec.version>1.14</commons-codec.version>
     </properties>
 
     <dependencyManagement>
@@ -69,6 +72,17 @@
                 <groupId>com.fasterxml.jackson.core</groupId>
                 <artifactId>jackson-databind</artifactId>
                 <version>${jackson.core.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>org.aspectj</groupId>
+                <artifactId>aspectjweaver</artifactId>
+                <version>${aspectjweaver.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>commons-codec</groupId>
+                <artifactId>commons-codec</artifactId>
+                <version>${commons-codec.version}</version>
             </dependency>
 
             <dependency>
