--- spring-boot-mock-backend/src/main/java/com/oopsmails/springboot/mockbackend/model/logging/ClientType.java
@@ -1,5 +1,0 @@
-package com.oopsmails.springboot.mockbackend.model.logging;
-
-public enum ClientType {
-    POSTMAN
-}
